watermark-tool/python/requirements.txt
blind-watermark==0.4.4
opencv-python==4.11.0.86
numpy==2.3.2
Pillow==11.3.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pytest==8.2.0
pyinstaller==6.19.0
